% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_car1.R
\name{add_car1}
\alias{add_car1}
\title{Add the CAR(1) autocorrelation structure.}
\usage{
add_car1(x, response, gr_vars = c(".index", "series"))
}
\arguments{
\item{x}{A dataframe.}

\item{response}{The column of \code{x} containing the resonse variable.}

\item{gr_vars}{Grouping variables withing which to calculate lagged residuals.}
}
\value{
The input dataframe with the fitted values transformed to have the CAR(1)
autocorrelation structure.
}
\description{
Add the CAR(1) autocorrelation structure.
}
