% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_car1_err.R
\name{add_car1_err}
\alias{add_car1_err}
\title{Add simulated CAR(1) error to model}
\usage{
add_car1_err(x, car1 = TRUE, gr_vars = c(".index", "series"))
}
\arguments{
\item{x}{A dataframe of model predictions.}

\item{car1}{Logical. Apply CAR(1) filter?}

\item{gr_vars}{Grouping variables defining groups within which to apply the CAR(1) filter.}
}
\value{
A dataframe of model predictions with CAR(1) errors.
}
\description{
Add simulated CAR(1) error to model
}
\examples{
car1_input <- expand.grid(.index = 1:2, location = letters[1:2], rep = 1:200)
car1_input$`ar[1]` <- .7
car1_input$nu <- 1e3
car1_input$sigma <- 1
car1_input$.epred <- 0
car1_input$d_x <- 1
add_car1_err(car1_input, gr_vars = c(".index", "location"))
}
