% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pred_draws_car1.R
\name{add_pred_draws_car1}
\alias{add_pred_draws_car1}
\title{Generate predictions for CAR(1) model}
\usage{
add_pred_draws_car1(
  input,
  object,
  type = "epred",
  car1 = TRUE,
  draw_ids = NULL,
  ...
)
}
\arguments{
\item{input}{A dataframe for which to generate model predictions.}

\item{object}{A \code{brms} model object.}

\item{type}{For \code{type = epred}, draws are from the expectation of the posterior predictive;
for\code{type = prediction}, draws are from the posterior predictive.}

\item{car1}{Logical. Add CAR(1) errors?}

\item{draw_ids}{Draw IDs from model object. If NULL (the default), all draws are used.}

\item{...}{Argument passed on to \code{tidybayes::add_epred_draws()}}
}
\value{
A dataframe of the type generated by \code{tidybayes::add_epred_draws()}.
}
\description{
Generate predictions for CAR(1) model
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3,
   backend = "cmdstanr"
 )
add_pred_draws_car1(data, fit, car1 = FALSE, draw_ids = 1234)
}
