% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_resid_draws_car1.R
\name{add_resid_draws_car1}
\alias{add_resid_draws_car1}
\title{Generate residual draws from a \code{brms} model object}
\usage{
add_resid_draws_car1(input, object, yvar, ...)
}
\arguments{
\item{input}{A dataframe for which to generate model predictions/residuals.}

\item{object}{A \code{brms} model object.}

\item{yvar}{Name of the response variable used to generate residuals.}

\item{...}{Arguments passed on to \code{add_pred_draws_car1()}.}
}
\value{
A dataframe of the type generated by \code{tidybayes::add_epred_draws()}.
}
\description{
Generate residual draws from a \code{brms} model object
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
add_resid_draws_car1(data, fit, yvar = y, car1 = FALSE, draw_ids = 1234)
}
