% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stancode.R
\name{add_stancode}
\alias{add_stancode}
\title{Add Stan code to the generated quantities block}
\usage{
add_stancode(scode_raw, new_code)
}
\arguments{
\item{scode_raw}{Stan code generated by \code{brms::make_stancode()}.}

\item{new_code}{A character string containing Stan code to add to the generated quantities block}
}
\value{
A Stan program of class \code{brmsmodel}.
}
\description{
Add Stan code to the generated quantities block
}
\examples{
library("brms")
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data_car1.csv"))
scode <- make_stancode(bf(y ~ ar(time = x)), data)
add_stancode(scode, "\n  // add new code here:")
}
