% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_acf.R
\name{calc_acf}
\alias{calc_acf}
\title{Calculate residual autocorrelation}
\usage{
calc_acf(x, ..., gr_vars = c(".draw", "series"), cen_var = NULL)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{...}{Arguments passed to a \code{data.table} chain for subsetting the data.frame after calculating the lagged
residual/censoring vectors.}

\item{gr_vars}{Calculate lagged residual/censoring vector within these groups.}

\item{cen_var}{Name of the censoring variable to be (optionally) lagged.}
}
\value{
A \code{tibble} with a Spearman rank correlation coefficient for each residual draw.
}
\description{
Calculate residual autocorrelation
}
\examples{

res <- data.frame(
.draw =  1,
.residual = c(1:9, -1e2),
location = "a"
)
calc_acf(res, gr_vars = c(".draw", "location"))
}
