% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ll.R
\name{calc_ll}
\alias{calc_ll}
\title{Calculate the log-likelihood of the posterior predictive distribution}
\usage{
calc_ll(x, response, censored = NULL, cens = TRUE)
}
\arguments{
\item{x}{A dataframe of posterior predictions.}

\item{response}{The name of the response variable.}

\item{censored}{The name of the censoring indicator.}

\item{cens}{Logical. Is the response left-censored?}
}
\value{
A matrix of log-likelihoods.
}
\description{
Calculate the log-likelihood of the posterior predictive distribution
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
 ll_in <- add_pred_draws_car1(data, fit, draw_ids = 1:10, car1 = FALSE)
 calc_ll(ll_in, "y", "ycens")
}
