% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{extract_params}
\alias{extract_params}
\title{Extract parameters from a \code{brms} model}
\usage{
extract_params(object, car1 = TRUE, draw_ids = NULL)
}
\arguments{
\item{object}{A \code{brms} model object.}

\item{car1}{Logical. Extract CAR(1) parameter?}

\item{draw_ids}{Numeric draw ids from the \code{brms} model. If NULL (the default), all draws are used.}
}
\value{
A dataframe of draws.
}
\description{
Extract parameters from a \code{brms} model
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
 extract_params(fit, car1 = FALSE, draw_ids = 1)
}
