% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_resp.R
\name{extract_resp}
\alias{extract_resp}
\title{Extract formula variables from a \code{brmsfit} object}
\usage{
extract_resp(x)
}
\arguments{
\item{x}{A \code{brmsfit} object}
}
\value{
A list of extracted formula variables for use by other \code{bgamcar1} functions
}
\description{
Extract formula variables from a \code{brmsfit} object
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
extract_resp(fit)
}
