% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_car1.R
\name{filter_car1}
\alias{filter_car1}
\title{CAR(1) filter}
\usage{
filter_car1(x, phi, s)
}
\arguments{
\item{x}{A numeric vector to be filtered.}

\item{phi}{The CAR(1) coefficient.}

\item{s}{A numeric vector defining the time spacing between observations.}
}
\value{
A numeric vector.
}
\description{
CAR(1) filter
}
\examples{
x <- rnorm(100)
d_x = rep(1, length(x))
filter_car1(x, .5, s = d_x)
}
