% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_censored.R
\name{impute_censored}
\alias{impute_censored}
\title{Impute left-censored observations via posterior prediction.}
\usage{
impute_censored(x, input, yvar, ycens, predvar = ".prediction")
}
\arguments{
\item{x}{A dataframe of observations and posterior predictions.}

\item{input}{The input dataframe used to generate the model.}

\item{yvar}{Name of the column containing left-censored observations.}

\item{ycens}{Name of the column containing the censoring indicator.}

\item{predvar}{Name of prediction column for imputing censored observations.}
}
\value{
A dataframe containing the (possibly imputed) observations.
}
\description{
Impute left-censored observations via posterior prediction.
}
\examples{
x <- data.frame(
   x = 1:10,
   y = rnorm(10),
   .prediction = rnorm(10),
   cens = "left"
)
impute_censored(x, x, "y", "cens")
}
