% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_slope.R
\name{local_slope}
\alias{local_slope}
\title{Calculate the local slope of a smooth term using finite differences}
\usage{
local_slope(input, object, x_var, epsilon = 0.001, smooth, g_var = NULL, ...)
}
\arguments{
\item{input}{A dataframe for which to generate model predictions.}

\item{object}{A \code{brms} model object.}

\item{x_var}{The predictor associated with the selected smooth term.}

\item{epsilon}{The increment on which to generate differences.}

\item{smooth}{The smooth term for which to calculate local slopes. Passed on to \code{brms::posterior_smooths()}.}

\item{g_var}{An optional grouping variable for factor-smooth interactions.}

\item{...}{Additional arguments passed to \code{brms::posterior_smooths()}.}
}
\value{
A \code{tibble} containing the (tidy) output of \code{brms::posterior_smooth()} and the calculated slopes.
}
\description{
Calculate the local slope of a smooth term using finite differences
}
\examples{
library("brms")
seed <- 1
data_gam <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data_gam.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test_gam"),
   seed,
   bf(y ~ s(x0) + s(x1) + s(x2) + s(x3)),
   data_gam,
   car1 = FALSE,
   chains = 2
 )
local_slope(data_gam, fit, "x2", smooth = "s(x2)")
}
