% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_cv.R
\name{loo_cv}
\alias{loo_cv}
\title{Approximate leave-one-out cross-validation (LOO) using \code{loo::loo()}.}
\usage{
loo_cv(input, object, censoring = TRUE, ...)
}
\arguments{
\item{input}{The dataframe of observations used to generate the model.}

\item{object}{A \code{brms} model object.}

\item{censoring}{Logical. Is the response left-censored?}

\item{...}{Arguments passed on to \code{add_pred_draws_car1()}.}
}
\value{
An object of class \code{loo}.
}
\description{
Approximate leave-one-out cross-validation (LOO) using \code{loo::loo()}.
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3,
   backend = "cmdstanr"
 )
 loo_cv(data, fit, draw_ids = 1:3000, car1 = FALSE)
}
