% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_stancode.R
\name{modify_stancode}
\alias{modify_stancode}
\title{Modify stan code generated by \code{brms} to fit a CAR(1) model}
\usage{
modify_stancode(scode_raw, modify = "car1")
}
\arguments{
\item{scode_raw}{Stan code generated by \code{brms::make_stancode()}.}

\item{modify}{Specify which modification to make. The default (and currently the only option) is \code{"car1"},
which creates the CAR(1) error term.}
}
\value{
A Stan program of class \code{brmsmodel}.
}
\description{
\code{modify_stancode()} resets the lower bound on the \code{ar} parameter to zero. It is compatible with priors such as
\code{normal()} (check to make sure it makes the expected modifications before using).
}
\examples{
library("brms")
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data_car1.csv"))
scode <- make_stancode(bf(y ~ ar(time = x)), data)
modify_stancode(scode)
}
