% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_km_nada.R
\name{ppc_km_nada}
\alias{ppc_km_nada}
\title{Generate ECDFs for posterior predictive checks with left-censored data}
\usage{
ppc_km_nada(input, object, draw_ids = 1:200, car1 = TRUE, seed, ...)
}
\arguments{
\item{input}{A dataframe for which to generate model predictions.}

\item{object}{A \code{brms} model object.}

\item{draw_ids}{Draw IDs from model object.}

\item{car1}{Logical. Add CAR(1) errors?}

\item{seed}{Set a random seed.}

\item{...}{Arguments passed on to \code{add_pred_draws_car1()}.}
}
\value{
A dataframe containing ECDFs generated by \code{NADA::cenfit()} for observations and model predictions.
}
\description{
Generate ECDFs for posterior predictive checks with left-censored data
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3,
   backend = "cmdstanr"
 )
 ppc_km_nada(data, fit, draw_ids = 34, seed = seed, car1 = FALSE)
}
