% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrans.R
\name{retrans}
\alias{retrans}
\title{Back-tranform scaled, (possibly) logged model predictions}
\usage{
retrans(x, scale_var, log = TRUE, recensor = FALSE, lcl = 2e-04)
}
\arguments{
\item{x}{A vector.}

\item{scale_var}{A vector by which to rescale \code{x}.}

\item{log}{Logical. Back-transform predictions using \code{exp()}?}

\item{recensor}{Logical. Recensor (left-censor) after back-transforming?}

\item{lcl}{Left-censoring limit for recensor = TRUE.}
}
\value{
A vector of back-transformed predictions.
}
\description{
Back-tranform scaled, (possibly) logged model predictions
}
\examples{
x <- rnorm(10, 10)
x_trans <- scale(x)[,1]
retrans(x_trans, x, log = FALSE)
}
