% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_preds.R
\name{summarize_preds}
\alias{summarize_preds}
\title{Summarize model predictions}
\usage{
summarize_preds(x, y_var = lead, retrans = TRUE, ...)
}
\arguments{
\item{x}{A dataframe of model predictions.}

\item{y_var}{The untransformed response, used to backtransform predictions.}

\item{retrans}{Logical. Backtransform predictions?}

\item{...}{Arguments passed on to \code{retrans()}.}
}
\value{
A dataframe containing the summarized predictions.
}
\description{
Summarize model predictions
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
preds <- add_pred_draws_car1(data, fit, car1 = FALSE, draw_ids = 1234)
summarize_preds(preds, y_var = y, log = FALSE)
}
