% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_car1.R
\name{add_car1}
\alias{add_car1}
\title{Add the CAR(1) autocorrelation structure.}
\usage{
add_car1(x, response, gr_vars = c(".index", "series"))
}
\arguments{
\item{x}{A dataframe.}

\item{response}{The column of \code{x} containing the resonse variable.}

\item{gr_vars}{Grouping variables withing which to calculate lagged residuals.}
}
\value{
The input dataframe with the fitted values transformed to have the CAR(1)
autocorrelation structure.
}
\description{
Add the CAR(1) autocorrelation structure.
}
\examples{
data_ar <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data_ar.csv"))
data_ar$`ar[1]` <- .7
data_ar$.epred <- 0
add_car1(data_ar, "y", gr_vars = "series")
}
