% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_model.R
\name{fit_stan_model}
\alias{fit_stan_model}
\title{Fit a stan model using (modified) code and data generated by brms}
\usage{
fit_stan_model(
  file,
  seed,
  bform,
  bdata,
  bpriors = NULL,
  car1 = TRUE,
  sample_prior = "no",
  knots = NULL,
  d_x = NULL,
  family = student(),
  backend = "rstan",
  overwrite = FALSE,
  var_car1 = NULL,
  var_xcens = NULL,
  cens_ind = NULL,
  lcl = NULL,
  lower_bound = NULL,
  stancode = NULL,
  ...
)
}
\arguments{
\item{file}{Load/save csvs to this filepath and name.}

\item{seed}{Random seed passed to \code{rstan::stan()}.}

\item{bform}{A \code{brms} formula. N.B., transformations of variables within the formula may not be picked up by
post-processing functions; it's generally better to create new, transformed variables.}

\item{bdata}{A data frame used to fit the model.}

\item{bpriors}{Priors specified by \code{brms} functions.}

\item{car1}{Logical. Generate CAR(1) errors?}

\item{sample_prior}{Passed on to \code{brms::brm()}.}

\item{knots}{Passed on to \code{brms::brm()}.}

\item{d_x}{A vector representing the spacing in time of observations in each series,
equal to zero at the first timestep. If \code{NULL} (the default), \code{d_x} is drawn from the dataframe \code{bdata}.}

\item{family}{A \code{brmsfamily} object. Note that some post-processing functions assume a student-t likelihood.}

\item{backend}{Run Stan's algorithms using \code{rstan} or \code{cmdstanr}.}

\item{overwrite}{Overwrite an exising model stored as CSVs? Defaults to \code{FALSE}.}

\item{var_car1}{For multivariate models, specify which variable to generate CAR(1) errors for.}

\item{var_xcens}{A character vector containing the names of variables with left-censoring; the order should
correspond to that of \code{lcl} and \code{cens_ind}.}

\item{cens_ind}{A character vector containing the names of left-censoring indicator variables corresponding
to the contents of \code{var_xcens}. (Indicator variables should take the form of character vectors with \code{"left"} for left-censored and
\code{"none"} for observed.)}

\item{lcl}{A numeric vector of censoring limits for left-censored variables. Alternatively, a list of numeric vectors with censoring limits for
each variable and observation; the order should correspond to that of \code{lcl} and \code{cens_ind}.}

\item{lower_bound}{An optional lower bound for left-censored parameters. The default is no lower bound.}

\item{stancode}{A named character vector of length one, where the name is the model block to modify
and the value is the additional Stan code. This has only been validated for the generated quantities block.}

\item{...}{Passed on to \code{rstan::stan()} or \code{cmdstanr::sample()}, depending on the backend selected.}
}
\value{
A \code{brms} model object fitted with \code{rstan} or \code{cmdstanr}.
}
\description{
Fit a stan model using (modified) code and data generated by brms
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
  paste0(system.file("extdata", package = "bgamcar1"), "/test"),
  seed,
  bf(y | cens(ycens, y2 = y2) ~ 1),
  data,
  prior(normal(0, 1), class = Intercept),
  car1 = FALSE,
  save_warmup = FALSE,
  chains = 3,
  backend = "cmdstanr"
)
}
