% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stan_model.R
\name{fit_stan_model}
\alias{fit_stan_model}
\title{Fit a stan model using (modified) code and data generated by brms}
\usage{
fit_stan_model(
  file,
  seed,
  bform,
  bdata,
  bpriors = NULL,
  car1 = TRUE,
  sample_prior = "no",
  knots = NULL,
  d_x = NULL,
  ...
)
}
\arguments{
\item{file}{Load/save csvs to this filepath and name.}

\item{seed}{Random seed passed to \code{rstan::stan()}.}

\item{bform}{A \code{brms} formula/}

\item{bdata}{A data frame used to fit the model.}

\item{bpriors}{Priors specified by \code{brms} functions.}

\item{car1}{Logical. Generate CAR(1) errors?}

\item{sample_prior}{Passed on to \code{brms::brm()}.}

\item{knots}{Passed on to \code{brms::brm()}.}

\item{d_x}{A vector representing the spacing in time of observations in each series,
equal to zero at the first timestep. If \code{NULL} (the default), \code{d_x} is drawn from the dataframe \code{bdata}.}

\item{...}{Passed on to \code{rstan::stan()}.}
}
\value{
A \code{brms} model object fitted with \code{rstan}.
}
\description{
Fit a stan model using (modified) code and data generated by brms
}
\examples{
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
  paste0(system.file("extdata", package = "bgamcar1"), "/test"),
  seed,
  bf(y | cens(ycens, y2 = y2) ~ 1),
  data,
  prior(normal(0, 1), class = Intercept),
  car1 = FALSE,
  save_warmup = FALSE,
  chains = 3
)
}
