% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_stancode.R
\name{modify_stancode}
\alias{modify_stancode}
\title{Modify stan code generated by \code{brms} to fit a CAR(1) model}
\usage{
modify_stancode(scode_raw)
}
\arguments{
\item{scode_raw}{Stan code generated by \code{brms::make_stancode()}.}
}
\value{
A Stan program of class \code{brmsmodel}.
}
\description{
\code{modify_stancode()} resets the lower bound on the \code{ar} parameter to zero. It is compatible with priors such as
\code{normal()} (check to make sure it makes the expected modifications before using).
}
\examples{
seed <- 1
library("brms")
seed <- 1
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data.csv"))
fit <- fit_stan_model(
   paste0(system.file("extdata", package = "bgamcar1"), "/test"),
   seed,
   bf(y | cens(ycens, y2 = y2) ~ 1),
   data,
   prior(normal(0, 1), class = Intercept),
   car1 = FALSE,
   save_warmup = FALSE,
   chains = 3
 )
modify_stancode(fit$model)
}
