% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_standata.R
\name{modify_standata}
\alias{modify_standata}
\title{Modify Stan data to include left-censoring limits on predictor variables.}
\usage{
modify_standata(sdata, data, lcl, var_xcens, cens_ind)
}
\arguments{
\item{sdata}{A list generated by \code{brms::make_standata()}}

\item{data}{A dataframe containing the model variables.}

\item{lcl}{A numeric vector of censoring limits for left-censored variables. Alternatively, a list of numeric vectors with censoring limits for
each variable and observation; the order should correspond to that of \code{lcl} and \code{cens_ind}.}

\item{var_xcens}{A character vector containing the names of predictor variables with left-censoring; the order should
correspond to that of \code{lcl} and \code{cens_ind}.}

\item{cens_ind}{A character vector containing the names of left-censoring indicators variables corresponding
to the contents of \code{var_xcens}.}
}
\value{
A list that can be passed to \code{rstan::stan()} or \code{cmdstanr::cmdstan_model}.
}
\description{
Modify Stan data to include left-censoring limits on predictor variables.
}
\examples{
library("brms")
data <- read.csv(paste0(system.file("extdata", package = "bgamcar1"), "/data_gam.csv"))
data$f_cens <- ifelse(data$f < 5, "left", "none")
sdata <- make_standata(bf(y ~ f), data)
sdata_modified <- modify_standata(sdata, data, lcl = 5, var_xcens = "f", cens_ind = "f_cens")
str(sdata_modified)
}
